fix: update template settings in menu
