+++ apps/journeys-admin/src/components/Editor/EditToolbar/Menu/TemplateSettingsDialog/useTemplateSettingsForm.ts
@@ -1,0 +1,11 @@
+import { FormikContextType, useFormikContext } from 'formik'
+
+import { JourneyUpdateInput } from '../../../../../../__generated__/globalTypes'
+
+export interface TemplateSettingsFormValues extends JourneyUpdateInput {
+  featured: boolean
+}
+
+export function useTemplateSettingsForm(): FormikContextType<TemplateSettingsFormValues> {
+  return useFormikContext<TemplateSettingsFormValues>()
+}
