--- apps/journeys-admin/src/components/Editor/EditToolbar/Menu/TemplateSettingsForm/index.ts
@@ -1,2 +1,0 @@
-export { TemplateSettingsForm } from './TemplateSettingsForm'
-export type { TemplateSettingsFormValues } from './TemplateSettingsForm'
